# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Easy Race Timer API
  version: 0.1.0
  description: |
    Early draft OpenAPI spec (partial). Covers competitions and entries endpoints.
servers:
  - url: https://api.example.com
    description: Production (placeholder)
  - url: http://localhost:3000
    description: Local development
paths:
...
  /competitors:
    get:
      summary: List competitors
      parameters:
        - in: header
          name: x-org-id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    post:
      summary: Create competitor
      parameters:
        - in: header
          name: x-org-id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
  /competitors/{id}:
    get:
      summary: Competitor detail
      parameters:
        - in: header
          name: x-org-id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  responses:
...
    NotFound:
      description: Resource not found
...